videostream/backend/requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
Django==4.0.4
djangorestframework==3.13.1
fonttools==4.33.3
idna==3.3
install==1.3.5
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sqlparse==0.4.2
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
